micro-services/redis-token/requirements.txt
grpcio==1.26.0
grpcio-tools==1.26.0
protobuf==3.11.2
python-dotenv==0.10.3
redis==3.3.11
six==1.13.0
